Fixed bug where id->misc->defines wasn't left behind properly after a top level RXML parse if there had been another RXML parse pass earlier in the request. This could show as a problem with missing javascript popups from the Javascript Support module (see [bug 3265]).
This has possible compatibility impact in case a third party module
counts on that the id->misc->defines from the first and not the last
(non-nested) RXML parse pass is present at the end of the request.
Since several non-nested RXML passes are uncommon it's probably a rare
problem, though.

Also improved the request trace a bit to show what is being done with
id->misc->defines.

Rev: server/base_server/rxml.pike:1.327