Better handling of RXML errors in the content of <append>.
Use RXML.t_string->encode(x) instead of (string)x in several places
where any value is converted to string. That gives RXML parse errors
instead of internal server errors if the type is incompatible so that
the cast fails.

Pass the frame for the <insert> tag to the insert plugin callbacks.

<insert variable="..."> now converts an array value to a NUL separated
string if it's inserted in a text context.

<emit source="values" from-scope="foo"> no longer trigs an internal
server error if some variable in the scope has a non-string|int|float
value.

Rev: server/modules/tags/rxmltags.pike:1.487