Fixed security issues where arbitrary databases in the Roxen server could be reached through the sql tags.
COMPATIBILITY/UPGRADE NOTE: Access is disabled by default, meaning
that essentially all sql tags that don't use the default database will
not work until the "Allowed databases" has been configured to local
needs. The module logs connection attempts to blocked databases in the
event log, so it is possible to see that way which ones that need to
be let through.

Also updated module docs a bit and did a small fix in the error
handling.

Rev: server/modules/database/sqltag.pike:1.112