Only create multiple module instances when explicitly asked for it in silent_modules.
Previously mentioning an already loaded module like "my_module" in
silent_modules caused a new instance to be added regardless whether
my_module was already loaded or not. With this patch it's necessary to
specify e.g. "my_module#1" to get a second instance.

Compatibility note: This change is not strictly compatible, but the
impact ought to be very small in practice.

Rev: server/site_templates/common.pike:1.29