Fixed missing propagation of state update events in the cache tag.
This bug could cause p-code updates to not be saved when there is a
nested update inside a cache hit in an existing persistently saved
cache entry. The nested update must affect the result p-code for this
bug to have effect - typically it's adding a new entry in a nested
cache tag that saves the cache persistently.

This is a regression from RXML.pmod/module.pmod rev 1.392, i.e. since
Roxen 4.5. In cases with nested persistent cache tags, it looks like
there have been a severe performance degradation since then, because
only the first cache entry in such caches got saved.

Creds to Wellhardh for finding this.

Rev: server/modules/tags/rxmltags.pike:1.639