<insert#href/>: Improved support for SOAP. Fixes some of [bug 6523].
This adds the attribute "soap-method", which causes the SOAPACTION
header in the request to be set, the request method be set to POST,
and any contained xml data to be sent.

Also unifies some more argument parsing between the threaded
and non-threaded implementations.

Rev: server/modules/tags/additional_rxml.pike:1.61