LogFormat: Improve the default log format pattern for events.
Add $ip-number, $user, $response and $length to the event log format.
They all (except for $length) expand to the same as the old hardcoded
value if they aren't available, so this should have a minimal effect
on logging for existing log facilities. It does however allow for
adding corresponding entries in the info mapping for new facilities,
without needing to add special case log format patterns.

Prepares for fixing [bug 6664].