Start: Add some extra stuff to the module path if existing.
The FSGC needs the Filesystem.Monitor module, which currently
(Pike 7.8) is provided by the feedimport module. As the
package system isn't yet in use in dists, it needs to be added
to the module path by the start-script. When the package system
is in use this is done when the package directories are added
by the roxenloader.

Fixes some more of [bug 6572].