FTP: Fixed SSL data connection.
The SSL/TLS handshaking for the data connection now seems to work.

Also adds compatibility with the old AUTH modes "SSL", "SSL-P" and "TLS-P".

Note that the old "AUTH SSL" and "AUTH SSL-P" modes imply SSL/TLS for the
data connection in active mode, but not in passive mode...

Verified to interoperate with the Linux ftp client version 0.17.