DBManager: Backup system fixes.
Don't attempt to use the old MySQL BACKUP and RESTORE commands with
MySQL 5.5 or later.

The automatic backup scheduler will now use the "mysqldump" method
for any databases where the "backup" method is selected, when the
internal MySQL server is version 5.5 or later.

Note that it is not possible to restore old backups made with the
BACKUP command with MySQL 5.5 or later. Any such backups should
be replaced with backups made with the mysqldump method prior to
upgrading the MySQL server.

Fixes some of [bug 6126], cf also [bug 7088].