CGI: Protect against bash(1) vulnerability.
bash(1) for some reason interprets environment variables with values
starting with "() {" as function definitions. These aliases are added
before any other processing.

In addition to the above, most versions of bash(1) released prior to
2014-09-24 had a bug, where any code after the function definition
was executed immediately.

This patch disables such functions by prefixing the value with a
single space.