The debug tag now takes content
If used with the werror attribute content in the debug tag will be printed to the logfile.

If the werror attribute is given the result from all eventual other attributes will be written to the log file and the result will be flushed, i.e the tag will not output anything.