Attempt to use protocol cache for certain authenticated resources as well.
All authenticated requests start out flagged as NO_PROTO_CACHE(), but if the
processing later decides it's cacheable after all we store it in a separate
cache compartment dedicated to data that looks like its protected but in
reality is available to everyone. Future authenticated requests may search
this compartment and get cache hits.