HTTPClient.pmod: Changed the method signatures so that the first argument is the URL to fetch. The other arguments aren't always needed so it's unnecessary to instantiate an object if it's not needed.
Also adds the host header to the request if it's not explicitly set already. This is due to a bug in Protocols.HTTP.Session where the port isn't set in the host header if the port isn't the standard port. A bug fix for that is checked in in Pike 8.1 not not yet in 8.0.