git-rxnpatch: Generate release sync clusters for all releases.
Previously a release sync cluster was only generated for the latest
release if there were patches for it.

Fixes issue where the meta cluster for Roxen 5.4.94-r3 contained
sync patches intended for Roxen 5.4.88-r2. This was due to the
sync clusters 5.4.88-5.4.94 not having been generated, and the
sync patches having ended up in the generic cluster.

Fixes [IS-26].