DBManager: Reduce critical zones holding sq_cache_lock().
sq_cache_{get,set}() now use sq_cache_lock() themselves instead
of having it be the callers responsibility. This makes it possible
to reduce the critical zones and to move out some of the time-
consuming stuff (notably the testing of the validity of cached
SQL connections) from the critical zones.

Fixes [WS-28].