Fetch Roxen news with AJAX.
Previously Roxen news was fetched async on the first page load (if no data was found in the cache) and a hard reload of the page was made after 4 seconds with the hope that the news had been fetched and cached during this period of time.
Butt ugly solution which is nothing but a memory now.