DBManager: Attempt to survive MySQL/MariaDB errors 185 and 190.
Some versions of MySQL and/or MariaDB apparently don't always
shut down in a forward-compatible way. Eg when upgrading from
MySQL 5.0.91 (ie Roxen 5.4.88) to MariaDB 10.1.12 (ie Roxen
6.1.200), may cause the table repair code to fail with errors
like "Incompatible key or row definition between the MariaDB
.frm file and the information in the storage engine. You have
to dump and restore the table to fix this".

If the repair code fails with the above error message, we now
retry the repair with the additional option USE_FRM as per
https://stackoverflow.com/questions/26871926/mariadb-mysql-database-table-corrupt-cant-repair .

Potential fix for [WS-46].