DBManager: Keep track of backups in progress.
Backups in progress now have a temporary entry in the
db_backups table (indicated as a backup of the table "").
The temporary entry is replaced by entries for the
individual tables when the backup completes.

This means that we no longer lose track of backups that are
in progress when the server restarts.

Fixes some of [WS-52].