Added "sticky" class.
Implemented in some files. This should be implemented where applicable.

Note: At the moment only Safari handles `position: sticky` on <thead>/<tr> tags. But since this doesn't break anything if it's not implemented in the browser it's safe to add. Chrome and Firefox will probably add support for sticky thead/tr elements in a not too distant future.