New feature: Modue Hot Reload.
If the server is started with `--debug` or `--module-debug` (implied with `--once`) you can call:

  ```
  void start() {
    my_configuration()->module_hot_reload(this);
  }
  ```

which will reload the module when the file is changed, without having to click reload in the admin interface.

If the server isn't started with any of the flags above the function call will have no effect.