Testsuite [WebDAV]: Fix handling of IPv4 ANY.
In Pike IPv4 ANY typically is represented as the zero ip-adress.
In Standards.URI however, setting the host field to zero is the
same as setting it to the empty string. To ensure the round-trip,
represent IPv4 ANY with the ip-adress "0.0.0.0" in URLs.

Fixes testsuite background failures "Invalid address ''".

Fixes [WS-314].