HTTP: Remove kludge working around bug in Netscape 4.7.
Netscape 4.7 (1999-09-30) didn't trust zero content-lengths when
using keep-alive, so we forced a Connection: Close in that case.

Netscape 4.7 is long obsolete, and all browsers released after
~2000 should have proper handling of Connection: Keep-Alive, and
the kludge was much too wide, so the time has come to remove it.

As a side-effect this fixes the WebDAV testsuite failing
due to too many TIME_WAIT sockets on Solaris 11.

Fixes [WS-315].