StartTLSProtocol: Changed ssl_keys to contain names of key pairs.
This changes ssl_keys from being a set of keypair identifier numbers
to a set of keypair names. This should make updating certificates
on the fly (with eg Let's Encrypt) behave more like users expect.

Previously if the certificate switched to a new key it was regarded
as a new certificate and manual configuration was needed. Now the
new certificate and key will most likely have the same name, and
will thus be used automatically.

Fixes [WS-580].