StartTLSProtocol: Fix dual use of filter_preferred_suites() arguments.
filter_preferred_suites() is used as a Variable.Variable change callback,
in which case it will be called with a Variable.Variable as the first
argument. This conflicted with the filter_preferred_suites() as it
expected a SSL.SSLContext as the first argument. This caused errors
like "Calling the NULL value.".

Move the SSL.SSLContext to the second argument.

Fixes [WS-618].