SQLKey: Fix wrapping of plain functions.
In some circumstances, the reference to the SQLKey is held only by
the current frame. Returning a function from the wrapped object may
then cause the SQLKey to lose all its references and then be
destructed before the returned function has been called, and during
the boot phase destruction of SQLKeys may cause destruction of the
wrapped object (and thus destruction of the returned function). We
now instead return a trampoline that holds a reference to the SQLKey.

Fixes [EP-2227].